Test update events with primitive pk types; split test-client into files
